// src/types/supabase.ts
export type Database = {
    // Allows to automatically instantiate createClient with right options
    // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
    __InternalSupabase: {
        PostgrestVersion: "14.1"
    }
    public: {
        Tables: {
// ...
            certifications: {
                Row: {
                    certificate_path: string | null
                    created_at: string | null
                    display_order: number | null
                    id: string
                    language: string
                    organization: string | null
                    title: string
                    type: string
                    year: string
                }
                Insert: {
                    certificate_path?: string | null
                    created_at?: string | null
                    display_order?: number | null
                    id?: string
                    language?: string
                    organization?: string | null
                    title: string
                    type: string
                    year: string
                }
                Update: {
                    certificate_path?: string | null
                    created_at?: string | null
                    display_order?: number | null
                    id?: string
                    language?: string
                    organization?: string | null
                    title?: string
                    type?: string
                    year?: string
                }
                Relationships: []
            }
// ...
        }
        Views: {
            [_ in never]: never
        }
        Functions: {
            [_ in never]: never
        }
        Enums: {
            [_ in never]: never
        }
        CompositeTypes: {
            [_ in never]: never
        }
    }
}